Added helper methods for info profile in _navbar.html
